src/requirements.txt
scikit-image
pandas
numpy
matplotlib
torch
torchvision
pyyaml